requirements.txt
requests==2.32.0
reportportal-client==5.5.4
pytest==8.1.1
selenium==3.13.0
python_dotenv==0.16.0
Appium_Python_Client==0.28
pytest-xdist>=1.31
pytest-html>=3.0.0
pytest-rerunfailures>=9.1.1
pytest_reportportal==5.4.0
pillow>=6.2.0
tesults
loguru
imageio
questionary>=1.9.0
clear-screen>=0.1.14
prompt-toolkit==2.0.10
bs4==0.0.2
jsonschema==4.21.1
pycryptodome